add reducer
