+++ src/components/Header.js
@@ -1,20 +1,28 @@
 import React,{ useState } from 'react'
 import styled from 'styled-components'
-import MenuIcon from '@mui/icons-material/Menu'
-import ClearIcon from '@mui/icons-material/Clear'
+import MenuIcon from '@mui/icons-material/Menu';
+import ClearIcon from '@mui/icons-material/Clear';
+import { selectCars } from '../features/car/carSlice';
+import CloseIcon from '@material-ui/icons/Close'
+import {useSelector} from 'react-redux'
+
 
 
 function Header() {
   const[burgerStatus, setBurgerStatus]=useState(false);
+  const cars=useSelector(selectCars)  
+  console.log(cars);
   return (
     <Container>
     <a href="#">
       <img src="/images/tesla_logomi.png" alt="" height="70px " width="150px"/></a>
      <Menu>
-      <a href="#">Model S</a>
-      <a href="#">Model 3</a>
-      <a href="#">Model X</a>
-      <a href="#">Model Y</a>
+      {cars && cars.map((car, index)=>(
+          <a key={index} href="#">{car}</a>
+          
+      ))}
+          
+      
       </Menu>  
       <RightMenu>
         <a href="#">Shop</a>
@@ -25,13 +33,16 @@ function Header() {
       <CloseWrapper>
             <CustomClose onClick={()=>setBurgerStatus(false)}/>
       </CloseWrapper>
+      {cars && cars.map((car, index)=>(
+          <li key={index}><a href="#">{car}</a></li>
+          
+      ))}
         <li><a href="#">Existing Inventory</a></li>
         <li><a href="#">Used Inventory</a></li>
         <li><a href="#">Trade-In</a></li>
         <li><a href="#">Cybertrucks</a></li>
         <li><a href="#">Roadaster</a></li>
-        <li><a href="#">Existing Inventory</a></li>
-        <li><a href="#">Existing Inventory</a></li>
+        
       </BurgerNav>
     </Container>
   )
