+++ src/features/counter/carSlice/carSlice.js
@@ -1,15 +1,15 @@
-import { configureStore } from '@reduxjs/toolkit'
+// import { configureStore } from '@reduxjs/toolkit'
 
-const initialState={ 
-          cars:["Model S", "Model 3", "Model X", "Model Y"]
+// const initialState={ 
+//           cars:["Model S", "Model 3", "Model X", "Model Y"]
 
-}
+// }
 
-const carSlice = createSlice({
-          name:"car",
-          initialState,
-          reducer:{}
-})
- export const selectCars= state => state.car.cars
+// const carSlice = createSlice({
+//           name:"car",
+//           initialState,
+//           reducer:{}
+// })
+//  export const selectCars= state => state.car.cars
 
- export default carSlice.reducer
+//  export default carSlice.reducer
